Player Movement Fixed
Player movement was giving A LOT of grief and I have no idea why. Cinemachine was going wonky and staying extremely zoomed in. Player movement was MUCH faster in the horizontal component than the vertical component (in raw input terms). Things were bugging out. Re-started Unity and re-pulled from main and returned the stashed changes and everything is working fine now. I can only imagine this is a Super Mario Speedrun situation where solar radiation flipped a bit in my computer and caused everything to go weird.

As for actual changes, the player moves in the direction that the camera is facing. Added drag variable. Added boolean to check if the player is touching the ground, which is checked via Raycast. Added a function to make sure that the player's speed matches the playerSpeed function.